+++ private/XCTest/XCTestManager_TestsInterface-Protocol.h
@@ -7,7 +7,7 @@
 @class NSData, NSString;
 
 @protocol XCTestManager_TestsInterface
-- (void)_XCT_receivedAccessibilityNotification:(int)arg1 withPayload:(NSData *)arg2;
+- (void)_XCT_receivedAccessibilityNotification:(long long)arg1 withPayload:(NSData *)arg2;
 - (void)_XCT_applicationWithBundleID:(NSString *)arg1 didUpdatePID:(int)arg2 andState:(unsigned long long)arg3;
 @end
